- Item type now is stored as uint16 instead of uint8 (indexes greater than 255 are available now). - Fixed a bug with weapon reloading. - Damage to map tiles is now randomized using "dDeviation" property. - Second Kratos' patch for muton armor. - Minor code clean-up. - Version bump.
git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@871 80a73507-52ba-0310-bdc1-b9b4a1856fab

+++ src/global.h
@@ -214,7 +214,7 @@ struct MANDATA
 struct ITEMDATA
 {
     uint8 num;
-    uint8 type[100];
+    uint16 type[100];
     uint8 place[100];
     uint8 x[100];
     uint8 y[100];
